Alterando a variável int id por Long id
Alterando a variável int id por Long id na classe GrupoDeUsuario

+++ src/model/GrupoDeUsuario.java
@@ -5,19 +5,19 @@ package model;
  * @author alexandre
  */
 public class GrupoDeUsuario {
-    private int id;
+    private Long id;
     private String nomeDoGrupo;
 
-public GrupoDeUsuario(int id, String nomeDoGrupo) {
+public GrupoDeUsuario(Long id, String nomeDoGrupo) {
         this.id = id;
         this.nomeDoGrupo = nomeDoGrupo;
     }
 
-public int getId() {
+public Long getId() {
         return id;
     }
 
-public void setId(int id) {
+public void setId(Long id) {
         this.id = id;
     }
